djajgo

+++ Project1/base/views.py
@@ -1,10 +1,13 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from .models import Room, Topic
 from .forms import RoomForm
+
 # Create your views here.
 
 # rooms =[
@@ -14,6 +17,9 @@ from .forms import RoomForm
 # ]
 
 def loginPage(request):
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -27,6 +33,7 @@ def loginPage(request):
             return redirect('home')
         else:
             return render(request, 'Username or Password does not exit')
+        
     context = {}
     return render(request, 'base/login_register.html', context)
 
@@ -48,6 +55,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -55,12 +63,17 @@ def createRoom(request):
         if form.is_valid():
             form.save()
             return redirect('home')
+        
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user != room.host:
+        return HttpResponse('You do not have permission to change here!!')
+    
     if request.method == 'POST':
         form = RoomForm(request.POST, instance = room)
         if form.is_valid():
@@ -70,9 +83,14 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You do not have permission to change here!!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
+    
     return render(request, 'base/delete.html', {'obj':room})
